Rename our hidden internal test module
