+++ src/lib.rs
@@ -28,7 +28,7 @@
 //! # /*
 //! let steam_dir = steamlocate::SteamDir::locate()?;
 //! # */
-//! # let temp_steam_dir = steamlocate::tests::helpers::expect_test_env();
+//! # let temp_steam_dir = steamlocate::__private_tests::helpers::expect_test_env();
 //! # let steam_dir = temp_steam_dir.steam_dir();
 //! println!("Steam installation - {}", steam_dir.path().display());
 //! // ^^ prints something like `Steam installation - C:\Program Files (x86)\Steam`
@@ -40,7 +40,7 @@
 //! assert_eq!(garrys_mod.name.as_ref().unwrap(), "Garry's Mod");
 //! println!("{garrys_mod:#?}");
 //! // ^^ prints something like vv
-//! # Ok::<_, steamlocate::tests::TestError>(())
+//! # Ok::<_, steamlocate::__private_tests::TestError>(())
 //! ```
 //! ```ignore
 //! App {
@@ -61,7 +61,7 @@
 //! # /*
 //! let steam_dir = steamlocate::SteamDir::locate()?;
 //! # */
-//! # let temp_steam_dir = steamlocate::tests::helpers::expect_test_env();
+//! # let temp_steam_dir = steamlocate::__private_tests::helpers::expect_test_env();
 //! # let steam_dir = temp_steam_dir.steam_dir();
 //!
 //! for library in steam_dir.libraries()? {
@@ -73,7 +73,7 @@
 //!         println!("    App {} - {:?}", app.app_id, app.name);
 //!     }
 //! }
-//! # Ok::<_, steamlocate::tests::TestError>(())
+//! # Ok::<_, steamlocate::__private_tests::TestError>(())
 //! ```
 //!
 //! On my laptop this prints
@@ -106,7 +106,7 @@ pub mod shortcut;
 // NOTE: exposed publicly, so that we can use them in doctests
 /// Not part of the public API >:V
 #[doc(hidden)]
-pub mod tests; // TODO: rename this since it may leak out in compiler error messages
+pub mod __private_tests;
 
 use std::collections::HashMap;
 use std::fs;
@@ -149,7 +149,7 @@ pub struct ReadmeDoctests;
 /// # /*
 /// let steam_dir = SteamDir::locate()?;
 /// # */
-/// # let temp_steam_dir = steamlocate::tests::helpers::expect_test_env();
+/// # let temp_steam_dir = steamlocate::__private_tests::helpers::expect_test_env();
 /// # let steam_dir = temp_steam_dir.steam_dir();
 /// assert!(steam_dir.path().ends_with("Steam"));
 /// ```
@@ -180,7 +180,7 @@ impl SteamDir {
     ///
     /// # Example
     /// ```
-    /// # let temp_steam_dir = steamlocate::tests::helpers::expect_test_env();
+    /// # let temp_steam_dir = steamlocate::__private_tests::helpers::expect_test_env();
     /// # let steam_dir = temp_steam_dir.steam_dir();
     /// # /*
     /// let steam_dir = SteamDir::locate()?;
@@ -189,7 +189,7 @@ impl SteamDir {
     /// let (warframe, library) = steam_dir.find_app(WARFRAME)?.unwrap();
     /// assert_eq!(warframe.app_id, WARFRAME);
     /// assert!(library.app_ids().contains(&warframe.app_id));
-    /// # Ok::<_, steamlocate::tests::TestError>(())
+    /// # Ok::<_, steamlocate::__private_tests::TestError>(())
     /// ```
     pub fn find_app(&self, app_id: u32) -> Result<Option<(App, Library)>> {
         // Search for the `app_id` in each library
